+++ course-overview.js
@@ -285,8 +285,8 @@ var courseOverviewData = [
     "hours": 8,
     "outline": {
       "exists": true,
-      "modules": 1,
-      "lessons": 13
+      "modules": 2,
+      "lessons": 5
     },
     "syllabus": true,
     "source": {
@@ -533,8 +533,8 @@ var courseOverviewData = [
     "hours": 50,
     "outline": {
       "exists": true,
-      "modules": 7,
-      "lessons": 14
+      "modules": 8,
+      "lessons": 17
     },
     "syllabus": true,
     "source": {
@@ -542,8 +542,8 @@ var courseOverviewData = [
       "folder": "Course: Excel for Data Analysts",
       "modules": 10
     },
-    "coverage": 14,
-    "lessonsWithContent": 2,
+    "coverage": 0,
+    "lessonsWithContent": 0,
     "assets": {
       "lessons": 15,
       "slides": 16,
@@ -564,8 +564,8 @@ var courseOverviewData = [
     "hours": 50,
     "outline": {
       "exists": true,
-      "modules": 8,
-      "lessons": 17
+      "modules": 5,
+      "lessons": 14
     },
     "syllabus": true,
     "source": {
@@ -1030,7 +1030,7 @@ var courseOverviewData = [
     "outline": {
       "exists": true,
       "modules": 1,
-      "lessons": 7
+      "lessons": 13
     },
     "syllabus": true,
     "source": {
@@ -1154,7 +1154,7 @@ var courseOverviewData = [
     "outline": {
       "exists": true,
       "modules": 1,
-      "lessons": 10
+      "lessons": 7
     },
     "syllabus": true,
     "source": {
@@ -1277,8 +1277,8 @@ var courseOverviewData = [
     "hours": 80,
     "outline": {
       "exists": true,
-      "modules": 5,
-      "lessons": 11
+      "modules": 4,
+      "lessons": 10
     },
     "syllabus": true,
     "source": {
@@ -1286,10 +1286,10 @@ var courseOverviewData = [
       "folder": "Course: Linux Foundations",
       "modules": 4
     },
-    "coverage": 91,
+    "coverage": 100,
     "lessonsWithContent": 10,
     "assets": {
-      "lessons": 10,
+      "lessons": 11,
       "slides": 10,
       "quizzes": 10,
       "activities": 23,
@@ -1300,7 +1300,7 @@ var courseOverviewData = [
       "modRecaps": 3,
       "interactives": 0
     },
-    "totalAssets": 73
+    "totalAssets": 74
   },
   {
     "id": "macos-administration-fundamentals",
@@ -1586,9 +1586,9 @@ var courseOverviewData = [
     "name": "Python Coding Booster Intensive",
     "hours": 16,
     "outline": {
-      "exists": true,
-      "modules": 2,
-      "lessons": 5
+      "exists": false,
+      "modules": 0,
+      "lessons": 0
     },
     "syllabus": true,
     "source": {
@@ -1596,7 +1596,7 @@ var courseOverviewData = [
       "folder": "Course: Python Booster",
       "modules": 8
     },
-    "coverage": 0,
+    "coverage": null,
     "lessonsWithContent": 0,
     "assets": {
       "lessons": 0,
@@ -1772,9 +1772,9 @@ var courseOverviewData = [
     "name": "SQL Coding Booster Intensive",
     "hours": 8,
     "outline": {
-      "exists": false,
-      "modules": 0,
-      "lessons": 0
+      "exists": true,
+      "modules": 1,
+      "lessons": 10
     },
     "syllabus": true,
     "source": {
@@ -1782,7 +1782,7 @@ var courseOverviewData = [
       "folder": "Course: SQL Booster",
       "modules": 4
     },
-    "coverage": null,
+    "coverage": 0,
     "lessonsWithContent": 0,
     "assets": {
       "lessons": 0,
@@ -1834,17 +1834,17 @@ var courseOverviewData = [
     "name": "SQL Fundamentals for Operations",
     "hours": 20,
     "outline": {
-      "exists": true,
-      "modules": 5,
-      "lessons": 14
+      "exists": false,
+      "modules": 0,
+      "lessons": 0
     },
-    "syllabus": true,
+    "syllabus": false,
     "source": {
       "exists": true,
       "folder": "Course: SQL for Data",
       "modules": 8
     },
-    "coverage": 0,
+    "coverage": null,
     "lessonsWithContent": 0,
     "assets": {
       "lessons": 17,